add fragment recyclerview City
