+++ app/src/main/res/layout/activity_popular_city.xml
@@ -1,33 +1,70 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
+<android.support.v4.widget.DrawerLayout
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        android:layout_width="match_parent" android:layout_height="match_parent"
         tools:context=".popularcityactivity.PopularCityActivity">
-    <TextView
+    <FrameLayout
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/mostPopularCities"
-            android:textColor="@color/colorBlack"
-            android:textStyle="bold"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            android:layout_marginTop="20dp"
-            android:padding="10dp"
-            android:textSize="20dp"
-            app:layout_constraintHorizontal_bias="0.5"
-            android:id="@+id/tvPopularCity"/>
+            android:layout_height="match_parent"
+            android:background="@color/colorWhite">
+        <android.support.constraint.ConstraintLayout
+                android:layout_width="match_parent"
+                android:layout_height="match_parent">
+            <ImageView
+                    android:id="@+id/imgDrawer"
+                    android:layout_width="0dp"
+                    android:layout_height="50dp"
+                    app:layout_constraintLeft_toLeftOf="parent"
+                    app:layout_constraintTop_toTopOf="parent"
+                    app:layout_constraintRight_toLeftOf="@+id/tvName"
+                    android:contentDescription="@string/todo"
+                    android:src="@drawable/ic_menu_black_36dp"/>
+            <TextView
+                    android:id="@+id/tvName"
+                    android:layout_width="0dp"
+                    android:layout_height="50dp"
+                    app:layout_constraintLeft_toRightOf="@+id/imgDrawer"
+                    app:layout_constraintRight_toLeftOf="@+id/imgSearch"
+                    android:text="@string/todo"
+                    android:textAlignment="center"
+                    android:gravity="center"
+                    app:layout_constraintTop_toTopOf="parent"/>
+            <ImageView
+                    android:id="@+id/imgSearch"
+                    android:layout_width="0dp"
+                    android:layout_height="50dp"
+                    app:layout_constraintLeft_toRightOf="@+id/tvName"
+                    app:layout_constraintTop_toTopOf="parent"
+                    app:layout_constraintRight_toRightOf="parent"
+                    android:src="@drawable/ic_search_black_36dp"
+            />
+            <FrameLayout
+                    android:id="@+id/frameLayoutDrawer"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    app:layout_constraintTop_toBottomOf="@id/tvName"
+            />
 
-    <android.support.v7.widget.RecyclerView
-            android:id="@+id/recyclerViewPopularCity"
+        </android.support.constraint.ConstraintLayout>
+    </FrameLayout>
+
+    <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="0dp"
-            tools:layout_editor_absoluteX="0dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/tvPopularCity"/>
-</android.support.constraint.ConstraintLayout>
+            android:layout_height="match_parent"
+            android:layout_gravity="start"
+            android:orientation="vertical"
+            android:background="@color/colorWhite">
+        <TextView
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="@string/todo"/>
+        <TextView
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="@string/todo"/>
+
+    </LinearLayout>
+
+</android.support.v4.widget.DrawerLayout>
