+++ app/src/main/res/layout/fragment_popular_city.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.v7.widget.RecyclerView
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        android:id="@+id/recyclerViewPopularCity"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"/>
