+++ app/src/main/res/values/colors.xml
@@ -4,4 +4,5 @@
     <color name="colorPrimaryDark">#00574B</color>
     <color name="colorAccent">#D81B60</color>
     <color name="colorBlack">#070707</color>
+    <color name="colorWhite">#FFFFFF</color>
 </resources>
